+++ src/OpenApi/JaneOpenApi.php
@@ -27,6 +27,8 @@ use Jane\OpenApi\Guesser\Guess\ClassGuess;
 class JaneOpenApi extends ChainGenerator
 {
     public const VERSION = '4.x-dev';
+    public const CLIENT_PSR18 = 'psr18';
+    public const CLIENT_HTTPLUG = 'httplug';
 
     private $schemaParser;
 
@@ -117,6 +119,10 @@ class JaneOpenApi extends ChainGenerator
 
     public static function build(array $options = [])
     {
+        if ($options['client'] === self::CLIENT_HTTPLUG) {
+            @trigger_error(sprintf('Generating "%s" client is deprecated, use the "%s" in the "client" option', self::CLIENT_HTTPLUG, self::CLIENT_PSR18));
+        }
+
         $encoders = [
             new JsonEncoder(
                 new JsonEncode(),