+++ src/OpenApi/Tests/fixtures/client-psr18/expected/Normalizer/NormalizerFactory.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Jane\OpenApi\Tests\Expected\Normalizer;
+
+class NormalizerFactory
+{
+    public static function create()
+    {
+        $normalizers = array();
+        $normalizers[] = new \Symfony\Component\Serializer\Normalizer\ArrayDenormalizer();
+        return $normalizers;
+    }
+}