+++ src/OpenApiRuntime/Client/BaseEndpoint.php
@@ -2,7 +2,6 @@
 
 namespace Jane\OpenApiRuntime\Client;
 
-use Http\Message\StreamFactory;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Serializer\SerializerInterface;
 
@@ -14,7 +13,7 @@ abstract class BaseEndpoint implements Endpoint
 
     abstract public function getMethod(): string;
 
-    abstract public function getBody(SerializerInterface $serializer, StreamFactory $streamFactory = null): array;
+    abstract public function getBody(SerializerInterface $serializer, $streamFactory = null): array;
 
     abstract public function getUri(): string;